Rearranging code
