+++ src/CarbonAware.CLI/src/commands/emissions/EmissionsCommand.cs
@@ -1,3 +1,4 @@
+using CarbonAware.Aggregators.CarbonAware;
 using CarbonAware.CLI.Common;
 using System.CommandLine;
 using System.CommandLine.Invocation;
@@ -6,25 +7,27 @@ namespace CarbonAware.CLI.Commands;
 
 class EmissionsCommand : Command
 {
-    private static readonly Option _startTimeOption = new Option<string>("--start-time", "The start time of the emissions query");
-    private static readonly Command _observedCommand = new ObservedCommand();
-
     public EmissionsCommand() : base("emissions", "emissions keyword")
     {
-        Add(_observedCommand);
-        Add(_startTimeOption);
+        CommonOptions.AddAllOptionsToCommand(this);
         this.SetHandler(this.Run);
     }
 
-    private void Run(InvocationContext context)
+    private async Task Run(InvocationContext context)
     {
-        var startTime = context.ParseResult.GetValueForOption(_startTimeOption);
-        var isVerbose = context.ParseResult.GetValueForOption(CommonOptions.VerbosityOption);
-        context.Console.Out.Write($"Emissions: {startTime}");
-        if (isVerbose)
-        {
-            context.Console.WriteLine($"VERBOSITY IS ON!");
+        // Get aggregator via DI.
+        var serviceProvider = context.BindingContext.GetService(typeof(IServiceProvider)) as IServiceProvider ?? throw new NullReferenceException("ServiceProvider not found");
+        var aggregator = serviceProvider.GetService(typeof(ICarbonAwareAggregator)) as ICarbonAwareAggregator ?? throw new NullReferenceException("CarbonAwareAggregator not found");
 
+        // Get the arguments and options to build the parameters.
+        var location = context.ParseResult.GetValueForOption<string>(CommonOptions.RequiredLocationOption) ?? "";
+        var parameters = new CarbonAwareParametersBaseDTO() { MultipleLocations = new string[] { location } };
+
+        // Call the aggregator.
+        var results = await aggregator.GetEmissionsDataAsync(parameters);
+        foreach (var result in results)
+        {
+            context.Console.WriteLine(result.ToString());
         }
         context.ExitCode = 0;
     }
